Support classes with access to multiple traditions or casting attributes
Add Barbarian and Alchemist

Note that this change doesn't allow for consolidating the draconic sorcerers, as focus spell tradition is distinct from spellcasting tradition from the data, but not in reality. The questions thus erroneously suggest that a draconic sorcerer can use divine spells and arcane focus spells, for example.